# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SwarmLens Control Plane API
  version: 1.5.0
  description: |
    Canonical v1.5 control-plane contract for SwarmLens.
    Cluster-scoped endpoints are primary. Legacy single-cluster aliases are intentionally omitted.
servers:
  - url: /api/v1
...
paths:
...
  /clusters:
    get:
      tags: [Clusters]
      summary: List configured clusters
      responses:
        '200':
          description: Cluster list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterListResponse'
    post:
      tags: [Clusters]
      summary: Create a cluster
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClusterCreateRequest'
      responses:
        '201':
          description: Created cluster
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterItemResponse'
  /clusters/{clusterID}:
    get:
      tags: [Clusters]
      summary: Get cluster
      parameters:
        - $ref: '#/components/parameters/ClusterID'
      responses:
        '200':
          description: Cluster detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Clusters]
      summary: Update cluster
      parameters:
        - $ref: '#/components/parameters/ClusterID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClusterUpdateRequest'
      responses:
        '200':
          description: Updated cluster
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /clusters/{clusterID}/swarm:
    get:
      tags: [Operations]
      summary: Get cluster posture
      parameters:
        - $ref: '#/components/parameters/ClusterID'
      responses:
        '200':
          description: Swarm posture
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwarmInfoItemResponse'
  /clusters/{clusterID}/nodes:
    get:
      tags: [Inventory]
      summary: List nodes
      parameters:
        - $ref: '#/components/parameters/ClusterID'
      responses:
        '200':
          description: Nodes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeListResponse'
  /clusters/{clusterID}/nodes/{id}:
    get:
      tags: [Inventory]
      summary: Get node
      parameters:
        - $ref: '#/components/parameters/ClusterID'
        - $ref: '#/components/parameters/EntityID'
      responses:
        '200':
          description: Node detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /clusters/{clusterID}/nodes/{id}/drain:
    post:
      tags: [Actions]
      summary: Drain node scheduling
      parameters:
        - $ref: '#/components/parameters/ClusterID'
        - $ref: '#/components/parameters/EntityID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionReasonRequest'
      responses:
        '200':
          description: Node drain outcome
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionOutcomeItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /clusters/{clusterID}/nodes/{id}/activate:
    post:
      tags: [Actions]
      summary: Resume node scheduling
      parameters:
        - $ref: '#/components/parameters/ClusterID'
        - $ref: '#/components/parameters/EntityID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionReasonRequest'
      responses:
        '200':
          description: Node activate outcome
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionOutcomeItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /clusters/{clusterID}/stacks:
    get:
      tags: [Inventory]
      summary: List stacks
      parameters:
        - $ref: '#/components/parameters/ClusterID'
      responses:
        '200':
          description: Stacks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StackListResponse'
  /clusters/{clusterID}/stacks/{name}:
    get:
      tags: [Inventory]
      summary: Get stack
      parameters:
        - $ref: '#/components/parameters/ClusterID'
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Stack detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StackItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Actions]
      summary: Remove stack
      description: Dry-run only in the current milestone.
      parameters:
        - $ref: '#/components/parameters/ClusterID'
        - in: path
          name: name
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionReasonRequest'
      responses:
        '200':
          description: Stack remove outcome
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionOutcomeItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /clusters/{clusterID}/stacks/{name}/deploy:
    post:
      tags: [Actions]
      summary: Deploy stack
      description: Dry-run only in the current milestone.
      parameters:
        - $ref: '#/components/parameters/ClusterID'
        - in: path
          name: name
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StackDeployRequest'
      responses:
        '200':
          description: Stack deploy outcome
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionOutcomeItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /clusters/{clusterID}/services:
    get:
      tags: [Inventory]
      summary: List services
      parameters:
        - $ref: '#/components/parameters/ClusterID'
        - in: query
          name: stack
          schema:
            type: string
      responses:
        '200':
          description: Services
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceListResponse'
  /clusters/{clusterID}/services/{id}:
    get:
      tags: [Inventory]
      summary: Get service
      parameters:
        - $ref: '#/components/parameters/ClusterID'
        - $ref: '#/components/parameters/EntityID'
      responses:
        '200':
          description: Service detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /clusters/{clusterID}/services/{id}/scale:
    post:
      tags: [Actions]
      summary: Scale service
      parameters:
        - $ref: '#/components/parameters/ClusterID'
        - $ref: '#/components/parameters/EntityID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceScaleRequest'
      responses:
        '200':
          description: Service scale outcome
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionOutcomeItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /clusters/{clusterID}/services/{id}/restart:
    post:
      tags: [Actions]
      summary: Restart service
      parameters:
        - $ref: '#/components/parameters/ClusterID'
        - $ref: '#/components/parameters/EntityID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionReasonRequest'
      responses:
        '200':
          description: Service restart outcome
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionOutcomeItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /clusters/{clusterID}/services/{id}/update:
    post:
      tags: [Actions]
      summary: Update service spec
      parameters:
        - $ref: '#/components/parameters/ClusterID'
        - $ref: '#/components/parameters/EntityID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceUpdateRequest'
      responses:
        '200':
          description: Service update outcome
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionOutcomeItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /clusters/{clusterID}/services/{id}/rollback:
    post:
      tags: [Actions]
      summary: Roll back service to previous spec
      parameters:
        - $ref: '#/components/parameters/ClusterID'
        - $ref: '#/components/parameters/EntityID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionReasonRequest'
      responses:
        '200':
          description: Service rollback outcome
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionOutcomeItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /clusters/{clusterID}/tasks:
    get:
      tags: [Inventory]
      summary: List tasks
      parameters:
        - $ref: '#/components/parameters/ClusterID'
        - in: query
          name: service
          schema:
            type: string
        - in: query
          name: node
          schema:
            type: string
        - in: query
          name: state
          schema:
            $ref: '#/components/schemas/TaskState'
      responses:
        '200':
          description: Tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
  /clusters/{clusterID}/tasks/{id}:
    get:
      tags: [Inventory]
      summary: Get task
      parameters:
        - $ref: '#/components/parameters/ClusterID'
        - $ref: '#/components/parameters/EntityID'
      responses:
        '200':
          description: Task detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /clusters/{clusterID}/tasks/{id}/restart:
    post:
      tags: [Actions]
      summary: Restart task
      parameters:
        - $ref: '#/components/parameters/ClusterID'
        - $ref: '#/components/parameters/EntityID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionReasonRequest'
      responses:
        '200':
          description: Task restart outcome
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionOutcomeItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /clusters/{clusterID}/networks:
    get:
      tags: [Inventory]
      summary: List networks
      parameters:
        - $ref: '#/components/parameters/ClusterID'
      responses:
        '200':
          description: Networks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NetworkListResponse'
  /clusters/{clusterID}/volumes:
    get:
      tags: [Inventory]
      summary: List volumes
      parameters:
        - $ref: '#/components/parameters/ClusterID'
      responses:
        '200':
          description: Volumes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VolumeListResponse'
  /clusters/{clusterID}/secrets:
    get:
      tags: [Inventory]
      summary: List secrets
      parameters:
        - $ref: '#/components/parameters/ClusterID'
      responses:
        '200':
          description: Secrets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecretListResponse'
  /clusters/{clusterID}/configs:
    get:
      tags: [Inventory]
      summary: List configs
      parameters:
        - $ref: '#/components/parameters/ClusterID'
      responses:
        '200':
          description: Configs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigListResponse'
  /clusters/{clusterID}/events:
    get:
      tags: [Inventory]
      summary: List recent cluster events
      parameters:
        - $ref: '#/components/parameters/ClusterID'
        - in: query
          name: type
          schema:
            type: string
      responses:
        '200':
          description: Events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwarmEventListResponse'
  /clusters/{clusterID}/stream/events:
    get:
      tags: [Inventory]
      summary: Stream cluster events
      parameters:
        - $ref: '#/components/parameters/ClusterID'
      responses:
        '200':
          description: Event stream
          content:
            text/event-stream:
              schema:
                type: string
  /clusters/{clusterID}/diagnostics:
    get:
      tags: [Operations]
      summary: List diagnostics findings
      parameters:
        - $ref: '#/components/parameters/ClusterID'
        - in: query
          name: severity
          schema:
            $ref: '#/components/schemas/Severity'
      responses:
        '200':
          description: Diagnostics findings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindingListResponse'
  /clusters/{clusterID}/diagnostics/{id}:
    get:
      tags: [Operations]
      summary: Get diagnostics finding
      parameters:
        - $ref: '#/components/parameters/ClusterID'
        - $ref: '#/components/parameters/EntityID'
      responses:
        '200':
          description: Diagnostics finding
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindingItemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /clusters/{clusterID}/incidents:
    get:
      tags: [Incidents]
      summary: List incidents
      parameters:
        - $ref: '#/components/parameters/ClusterID'
      responses:
        '200':
          description: Incidents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentListResponse'
    post:
      tags: [Incidents]
      summary: Create incident
      parameters:
        - $ref: '#/components/parameters/ClusterID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentCreateRequest'
      responses:
        '201':
          description: Created incident
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentItemResponse'
...
components:
  parameters:
    ClusterID:
      in: path
      name: clusterID
      required: true
      schema:
        type: string
    EntityID:
      in: path
      name: id
      required: true
      schema:
        type: string
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
    cookieSession:
      type: apiKey
      in: cookie
      name: swarmlens_session
  schemas:
...
    Severity:
      type: string
      enum: [critical, high, medium, low, info]
...
    ClusterCreateRequest:
      type: object
      required: [name, dockerHost]
      properties:
        name:
          type: string
        dockerHost:
          type: string
        connectionMode:
          $ref: '#/components/schemas/ClusterConnectionMode'
        tlsEnabled:
          type: boolean
        certRef:
          type: string
        enabled:
          type: boolean
        default:
          type: boolean
...
    ClusterItemResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/Cluster'
    ClusterListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Cluster'
        meta:
          $ref: '#/components/schemas/ListMeta'
    SwarmInfoItemResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/SwarmInfo'
...